+++ src/components/Nav/index.ts
@@ -1,0 +1,2 @@
+export { default as NavButton } from '@/components/Nav/NavButton';
+export { default as NavListItem } from '@/components/Nav/NavListItem';
